add minimal example of writing/reading a zarr from dask, refactor
